move build_views_for_stream into masker
